Fix change detection for Lambda task actions
